Remove unused code related to QueryCache statistics
The querycache data was commented out in the template with this note:
"Not supported with Creole" - so it was not used anyway.
And call to getStorage() failed because it is "filesystem" cache
and tries to open non-existing dir "var/cache/database" in
realpath() call in "lib/xarigami/caching/storage/filesystem.php".